Add LED and more board tweaking.
Added another LED after realizing that STAT on the battery charger IC is
actually tri-state -- thanks, Adafruit!  This means we can now show when
the battery is charging and when it's full, wee!

Also moved bits and pieces around on the board.  Spread out the
stitching vias, moved vias out of the way of the silkscreen, etc etc.