Created an enclosure for the circuit board.
I whipped up a simple enclosure for the board in SolidWorks.  Some of
the features:

- has enough space under the board for the LiPo battery to be nestled
- locating pins to lock the board in place
- snap hooks for toolless assembly

I'm pretty sure we should be able to 3D print this, although the feature
sizes on the snap hooks is a little small, so I'll have to get a test
printing and see how it fares.