Final layout tweaks before locking rev A.
We've given the TPS61200 its own power ground, hopefully quashing as
much noise as possible off the main ground, keeping RF performance high.

I lied, and went back to adjust the stitching.  We should have our
stitching on a 100 mil grid consistently, with a few exceptions to deal
with component placement.

Also fixed a bug where we had a decoupling cap on a configuration pin
instead of a supply pin.