Tweaks t to the enclosure dimensions.
After getting prototypes made, there were a few design errors that
needed to be corrected:
- didn't include any tolerance for the locating pins
- board was too snug of a fit between snap hooks to easily seat
- edge fillets were too small and ended up more like chamfers
- didn't propagate the HDMI connector cut-out to the assembly parts

The board now has a 1mm clearance around the perimeter except for the
connector edge.  The fillet on the enclosure edges is also 1mm larger
which should actually provide a more rounded edge.

The locating pins are also 0.1mm smaller in radius, while the locating
pin holes are 0.1mm larger in radius.  This will make them easily to go
through the holes in the board, and actually fit each other when closing
the assembly.

Finally, while small, the cut extrude to smooth the edges of the HDMI
connector hole (because of where the split is) looked correct in the
assembly but wasn't originally propagated to the top and bottom pieces,
so it was never actually printed that way.